some babel/emit/error reporting fixes
This commit needs to be fixed up and documented -- not ready for merge
-- but seems to improve
some things regarding emits as well as having fixes for
externalTranspiler/babel support.

Because many tranpiles are synchronous, the main feature of these
changes is to make externalTranipilation synchronous so as not to fall
out of the event loop, but it also makes some changes to emit handling
and tranpile error handling/reporting.

It might help with issues such as #737 and perhaps even #765 (it
doesn't exhibit the Emit Failed symptoms even with noEmit: true).

For now, it doesn't close any issues. It's also behin master, so
certainly beware.